Made long conversion conistent. Improved test coverage. JIRA: MATH-224.
git-svn-id: https://svn.apache.org/repos/asf/commons/proper/math/trunk@791931 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/org/apache/commons/math/stat/descriptive/AggregateSummaryStatistics.java
@@ -238,7 +238,7 @@ public class AggregateSummaryStatistics implements StatisticalSummary,
             }
             sum += current.getSum();
             final double oldN = n;
-            final long curN = current.getN();
+            final double curN = current.getN();
             n += curN;
             final double meanDiff = current.getMean() - mean;
             mean = sum / n;
